chore: move printing of sample config file out of config.go (#11409)
